Move applyPwToInjectorChannels() to scheduler_fuel_controller
